requirements.txt
flask==3.0.2
beautifulsoup4==4.12.3
requests==2.31.0
pymupdf==1.23.23
openai
mistralai
google-generativeai
langchain==0.1.7
numpy
scikit-learn